Major cleanup: archive sprawl, update docs for distributed compute milestone
- Archive 1,600+ stale docs and showcase files (still in git history)
- Add .cursorignore to prevent IDE from indexing archives
- Update root docs (README, STATUS, DOCUMENTATION, QUICK_STATUS, QUICK_REFERENCE)
  to reflect Feb 9 cross-vendor distributed GPU compute validation
- Add toadstool-core, akida-driver backends, interpolation/linalg modules,
  cross-hardware parity tests, scientific computing shaders
- Add session handoff for distributed compute (NVIDIA + AMD bit-identical)
- Fix last production unwrap() in tensor.rs
- Net: -550K lines of archived sprawl, clean 6-file root doc structure

+++ crates/runtime/display/src/drm/device.rs
@@ -99,7 +99,7 @@ impl Device {
         )
         .map_err(|e| {
             tracing::error!("Failed to open {}: {}", path.display(), e);
-            DisplayError::OpenFailed(std::io::Error::from_raw_os_error(e.raw_os_error() as i32))
+            DisplayError::OpenFailed(std::io::Error::from_raw_os_error(e.raw_os_error()))
         })?;
 
         let fd = Arc::new(fd);
